MM-26: wip

+++ app/Models/Photo.php
@@ -11,52 +11,64 @@ use Illuminate\Support\Str;
 /**
  * Class Photo
  *
- * @property-read Model $photoable  Polymorphic relation to the owning model.
+ * @method photoable()
+ *
+ * @property \Illuminate\Database\Eloquent\Model $photoable
+ *
+ * @property int $id
  * @property string $path
  * @property bool $is_main
- * @property string $photoable_type
- * @property int $photoable_id
- * @property \Illuminate\Support\Carbon $created_at
- * @property \Illuminate\Support\Carbon $updated_at
+ * @property bool $is_public
+ *
+ *
  */
 class Photo extends Model
 {
     use HasFactory;
 
+    /**
+     * Get the attributes that should be cast.
+     *
+     * @return array<string, string>
+     */
+    protected function casts(): array
+    {
+        return [
+            'is_main'   => 'boolean',
+            'is_public' => 'boolean',
+        ];
+    }
+
     protected static function boot(): void
     {
         parent::boot();
 
-        static::deleting(function (Photo $photo) {
-            Storage::disk('public')->delete($photo->path);
+        static::deleting(function ($photo) {
+            if (Storage::disk('public')->exists($photo->path)) {
+                Storage::disk('public')->delete($photo->path);
+            }
         });
 
-        static::saving(function (Photo $photo) {
+        static::saving(function ($photo) {
             if ($photo->isDirty('path')) {
                 $photo->handlePhotoSaving();
             }
         });
     }
 
-    /**
-     * Define the polymorphic relationship.
-     *
-     * @return MorphTo
-     */
     public function photoable(): MorphTo
     {
         return $this->morphTo();
     }
 
-    /**
-     * Handle the logic for saving a photo.
-     */
     protected function handlePhotoSaving(): void
     {
-        $oldPhoto  = $this->getOriginal('path');
+        if (!$this->path) {
+            return;
+        }
+
         $directory = $this->photoable->getPhotoDirectory(); //@phpstan-ignore-line
 
-        // Generate new file name
         $newFileName = sprintf(
             '%s_%s.%s',
             Str::slug($this->photoable->getPhotoNamePrefix()), //@phpstan-ignore-line
@@ -64,18 +76,11 @@ class Photo extends Model
             pathinfo($this->path, PATHINFO_EXTENSION)
         );
 
-        // Define the new path
         $newFilePath = $directory . '/' . $newFileName;
 
-        // Move the file to the new location
-        Storage::disk('public')->move($this->path, $newFilePath);
-
-        // Update the path
-        $this->path = $newFilePath;
-
-        // Delete the old file, if it exists and is different
-        if ($oldPhoto && $oldPhoto !== $this->path) {
-            Storage::disk('public')->delete($oldPhoto);
+        if (Storage::disk('public')->exists($this->path)) {
+            Storage::disk('public')->move($this->path, $newFilePath);
+            $this->path = $newFilePath;
         }
     }
 }
